requirements.txt
Flask==2.2.5
line-bot-sdk==2.5.0
pillow==10.2.0
pytesseract==0.3.10